перенес некоторые модули в папку ui
