+++ niffler_e_2_e_tests_python/pages/login_page.py
@@ -3,8 +3,8 @@ from typing import TYPE_CHECKING
 import allure
 from configs import configs
 from pages.presentation_page import PresentationPage
-from utils.playwright_helper import PlaywrightHelper
-from utils.utils import get_join_url
+from tests_ui.utils.playwright_helper import PlaywrightHelper
+from tests_ui.utils.utils import get_join_url
 
 if TYPE_CHECKING:
     from playwright.sync_api import Page
