+++ niffler_e_2_e_tests_python/tests_ui/presentation/authorization/main/conftest.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
     from playwright.sync_api import Page
 
 
-pytest_plugins = ('tests.authorization.main.profile.conftest')
+pytest_plugins = ('tests_ui.presentation.authorization.main.profile.conftest')
 
 
 @pytest.fixture(scope='session')
