+++ niffler_e_2_e_tests_python/tests_ui/presentation/authorization/main/profile/test_profile.py
