+++ niffler_e_2_e_tests_python/tests_ui/presentation/authorization/test_authorization.py
@@ -4,7 +4,7 @@ import allure
 import pytest
 from configs import configs
 from faker import Faker
-from tests_ui.authorization.enums import ErrorAuthorization
+from tests_ui.presentation.authorization.enums import ErrorAuthorization
 
 if TYPE_CHECKING:
     from pages.login_page import LoginPage
