+++ niffler_e_2_e_tests_python/tests_ui/presentation/registration/conftest.py
@@ -6,7 +6,7 @@ from pages.register_page import RegisterPage
 if TYPE_CHECKING:
     from playwright.sync_api import Page
 
-pytest_plugins = ('tests.authorization.conftest', 'tests.authorization.main.conftest')
+pytest_plugins = ('tests_ui.presentation.authorization.conftest', 'tests_ui.presentation.authorization.main.conftest')
 
 
 @pytest.fixture(scope='session')
