+++ niffler_e_2_e_tests_python/tests_ui/presentation/registration/test_registration.py
